+++ tests/der_constructed.rs
@@ -9,7 +9,7 @@ use nom::{Err, Needed};
 use test_case::test_case;
 
 #[test_case(&hex!("a2 05 02 03 01 00 01"), Ok(0x10001) ; "tag ok")]
-#[test_case(&hex!("a2 80 02 03 01 00 01 00 00"), Err(BerError::IndefiniteLengthUnexpected) ; "indefinite tag ok")]
+#[test_case(&hex!("a2 80 02 03 01 00 01 00 00"), Err(BerError::DerConstraintFailed(DerConstraint::IndefiniteLength)) ; "indefinite tag ok")]
 #[test_case(&hex!("a3 05 02 03 01 00 01"), Err(BerError::unexpected_tag(Some(Tag(2)), Tag(3))) ; "invalid tag")]
 #[test_case(&hex!("22 05 02 03 01 00 01"), Err(BerError::unexpected_class(None, Class::Universal)) ; "invalid class")]
 #[test_case(&hex!("82 05 02 03 01 00 01"), Err(BerError::ConstructExpected) ; "construct expected")]
@@ -57,7 +57,7 @@ fn tc_der_tagged_implicit_g(i: &[u8], out: Result<u32, BerError>) {
 #[test_case(&hex!("30 0a 02 03 01 00 01 02 03 01 00 00"), Ok(&[0x10001, 0x10000]) ; "seq ok")]
 #[test_case(&hex!("30 07 02 03 01 00 01 02 03 01"), Err(BerError::NomError(ErrorKind::Eof)) ; "incomplete")]
 #[test_case(&hex!("31 0a 02 03 01 00 01 02 03 01 00 00"), Err(BerError::unexpected_tag(Some(Tag::Sequence), Tag::Set)) ; "invalid tag")]
-#[test_case(&hex!("30 80 02 03 01 00 01 00 00"), Err(BerError::IndefiniteLengthUnexpected) ; "indefinite seq ok")]
+#[test_case(&hex!("30 80 02 03 01 00 01 00 00"), Err(BerError::DerConstraintFailed(DerConstraint::IndefiniteLength)) ; "indefinite seq ok")]
 fn tc_der_seq_of(i: &[u8], out: Result<&[u32], BerError>) {
     fn parser(i: &[u8]) -> BerResult {
         parse_der_sequence_of(parse_der_integer)(i)
@@ -85,7 +85,7 @@ fn tc_der_seq_of(i: &[u8], out: Result<&[u32], BerError>) {
 #[test_case(&hex!("30 0a 02 03 01 00 01 02 03 01 00 00"), Ok(&[0x10001, 0x10000]) ; "seq ok")]
 #[test_case(&hex!("30 07 02 03 01 00 01 02 01"), Err(Err::Incomplete(Needed::new(1))) ; "incomplete")]
 #[test_case(&hex!("31 0a 02 03 01 00 01 02 03 01 00 00"), Err(Err::Error(BerError::unexpected_tag(Some(Tag::Sequence), Tag::Set))) ; "invalid tag")]
-#[test_case(&hex!("30 80 02 03 01 00 01 00 00"), Err(Err::Error(BerError::IndefiniteLengthUnexpected)) ; "indefinite seq ok")]
+#[test_case(&hex!("30 80 02 03 01 00 01 00 00"), Err(Err::Error(BerError::DerConstraintFailed(DerConstraint::IndefiniteLength))) ; "indefinite seq ok")]
 fn tc_der_seq_defined(i: &[u8], out: Result<&[u32], Err<BerError>>) {
     fn parser(i: &[u8]) -> BerResult<DerObject> {
         parse_der_sequence_defined(map(
@@ -117,7 +117,7 @@ fn tc_der_seq_defined(i: &[u8], out: Result<&[u32], Err<BerError>>) {
 #[test_case(&hex!("31 0a 02 03 01 00 01 02 03 01 00 00"), Ok(&[0x10001, 0x10000]) ; "set ok")]
 #[test_case(&hex!("31 07 02 03 01 00 01 02 03 01"), Err(BerError::NomError(ErrorKind::Eof)) ; "incomplete")]
 #[test_case(&hex!("30 0a 02 03 01 00 01 02 03 01 00 00"), Err(BerError::unexpected_tag(Some(Tag::Set), Tag::Sequence)) ; "invalid tag")]
-#[test_case(&hex!("31 80 02 03 01 00 01 00 00"), Err(BerError::IndefiniteLengthUnexpected) ; "indefinite set ok")]
+#[test_case(&hex!("31 80 02 03 01 00 01 00 00"), Err(BerError::DerConstraintFailed(DerConstraint::IndefiniteLength)) ; "indefinite set ok")]
 fn tc_der_set_of(i: &[u8], out: Result<&[u32], BerError>) {
     fn parser(i: &[u8]) -> BerResult {
         parse_der_set_of(parse_der_integer)(i)
@@ -145,7 +145,7 @@ fn tc_der_set_of(i: &[u8], out: Result<&[u32], BerError>) {
 #[test_case(&hex!("31 0a 02 03 01 00 01 02 03 01 00 00"), Ok(&[0x10001, 0x10000]) ; "set ok")]
 #[test_case(&hex!("31 07 02 03 01 00 01 02 01"), Err(Err::Incomplete(Needed::new(1))) ; "incomplete")]
 #[test_case(&hex!("30 0a 02 03 01 00 01 02 03 01 00 00"), Err(Err::Error(BerError::unexpected_tag(Some(Tag::Set), Tag::Sequence))) ; "invalid tag")]
-#[test_case(&hex!("31 80 02 03 01 00 01 00 00"), Err(Err::Error(BerError::IndefiniteLengthUnexpected)) ; "indefinite set ok")]
+#[test_case(&hex!("31 80 02 03 01 00 01 00 00"), Err(Err::Error(BerError::DerConstraintFailed(DerConstraint::IndefiniteLength))) ; "indefinite set ok")]
 fn tc_der_set_defined(i: &[u8], out: Result<&[u32], Err<BerError>>) {
     fn parser(i: &[u8]) -> BerResult<DerObject> {
         parse_der_set_defined(map(